requirements.txt
pandas
streamlit-option-menu
pymongo
wordcloud